+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Replace `fd-slicer` module with internal transform stream
   * deps: http-errors@~1.8.0
     - Fix error creating objects in some environments
     - deps: inherits@2.0.4